+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add logging of uncaught asynchronous error logs to a hardcoded path in /tmp
 - Bug Fix: Drop diagnostics for files not opened yet in the editor instead of
   crashing
+- Add support for textDocument/references call. Only supports showing references
+  in files that have already been opened by the client.
 
 # 0.0.6